Add a ScannerResultBuilder implementation that streams results
StreamingScannerResultBuilder uses the Jackson Streaming API to
generate an OrtResult directly on disk. It writes out scan results
when they arrive; so results do not have to be kept in memory.

Add more tests to verify that the results produced can be read and
further processed using ORT's object model.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>