Scanner: Process incoming scan results one by one
Do not collect all results of a scan operation and write them to the
builder at the end, but pass them to the builder as soon as they
become available. This should reduce memory consumption, as the
single result objects can be subject to garbage collection after they
have been processed.

Use a channel to implement the sequential processing of scan results.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>