LocalScanner: Check for archived files if scan results are present
The LocalScanner supports both an archive and a storage. So far, if
results of a package are present in the storage, they are skipped,
independent of the presence of results in the archive. Add an additional
check that, in case of absence, downloads the source code and applies the
archiving process.

Signed-off-by: Korbinian Singhammer <external.Korbinian.Singhammer2@bosch.io>